Block 734838 * Next Difficulty Adjustment = 735839
Timestamp - Block 734838    ($38,390 USD)

1001 blocks to re-target 

Next Adjustment Date:  10 May 2022   Block 735839

Difficulty is a value used to show how hard is it to find a hash that will be lower than target defined by system. The Bitcoin network has a global block difficulty. Valid blocks must have a hash below this target. Mining pools also have a pool-specific share difficulty setting a lower limit for shares.

"The current Bitcoin mining difficulty is very high."

The difficulty is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

"With e-currency based on cryptographic proof, without the need to trust a third party middleman, money can be secure and transactions effortless."  - satoshi