Block 738169 * Next Difficulty Adjustment = 739871
Timestamp - Block 738169    ($28,800.42 USD)

1702 blocks to re-target 

Next Adjustment Date:   08 June 2022   @ Block 739871

Difficulty is a value used to show how hard is it to find a hash that will be lower than target defined by system.  The Bitcoin network has a global block difficulty.  Valid blocks must have a hash below this target.  Mining pools also have a pool-specific share difficulty setting a lower limit for shares.

The difficulty is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks.  At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find.  If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced.  If they took less than two weeks, the difficulty is increased.  The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find.  To find a block, the hash must be less than the target.  The hash is effectively a random number between 0 and 2**256-1.

" Sigh... why delete a wallet instead of moving it aside and keeping the old copy... just in case?  You should never delete a wallet.. juts sayin', just.. sayin' .  "