744288 ⚡️ Next Diff. Adjustment - Block 745919
Timestamp ⚡️ Block 744288  ($21,629)

1631 Blocks to retarget

Next Difficulty Adjustment -  20 JULY 2022  08:08:30 ET  @ Block 745919

The difficulty is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks.  At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find.  If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced.  If they took less than two weeks, the difficulty is increased.  The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find.  To find a block, the hash must be less than the target.  The hash is effectively a random number between 0 and 2**256-1.

" There will be transaction fees, so nodes will have an incentive to receive and include all the transactions they can.  Nodes will eventually be compensated by transaction fees alone when the total coins created hits the pre-determined ceiling. "