Timestamp: ₿lock 751743 ⌛ AUG 29, 2022
Reward Per Block	=  6.25+0.06822 BTC ($127,388.38) 

Avg. Transaction Fee	= 0.000037 BTC ($0.742) 0.000000099 BTC/byte

    Hashrate  	213.089 Ehash/s   (-6.94% in 24 hours)
    Difficulty	28.352 T  

Next retarget @ block 751968   (in 225 blocks ~ 1 days 10 hours)

The difficulty is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.