Timestamp: ₿lock 752,203 ⌛ SEPT 01, 2022
Difficulty is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Reward Per Block	6.25+0.1118 BTC ($128,067.9)

Difficulty   30.977 T   (+3.11% in 24 hours)
Hashrate   219.784 Ehash/s   (-6.27% in 24 hours)

Next retarget @ block 753984   (in 1781 blocks ~ 11 days 11 hours)