⚡ lnd:  01:55p.m ET September 03, 2022
The statistics are aggregated and calculated from multiple nodes within the Lightning Network. Due to the decentralized nature of the Lightning Network, the numbers observed are approximations and nodes that don't broadcast their state are NOT included***

CONNECT to 1ML public node:
0217890e3aad8d35bc054f43acc00084b25229ecff0ab68debd82883ad65ee8266@23.237.77.11:9735


BTC TIMECHAIN ~ Difficulty Adjustment:
Difficulty	=  30.977 T 

DIFFICULTY* ...is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 753984   (in 1548 blocks ~ 10 days 12 hours)