Ts: ₿lock 752,602 ⌛ Next retarget @ block 753984
Avg. Transaction Fee	=   0.00006 BTC   ($1.19)   0.00000015 BTC/byte
    Median Transaction Fee	=   0.000025 BTC   ($0.492)

Reward Per Block	=   6.25+0.06139 BTC   ($124,481.12) 
    Next halving @ block 840000   (in 87398 blocks ~ 595 days)

 Hashrate   =   242.821 Ehash/s   (+28.34% in 24 hours)
 Difficulty    =   30.977 T 

Difficulty is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 753984   (in 1382 blocks ~ 9 days 8 hours)