⌛ ₿lock 755546 ⚡ Next retarget @ block 756000
TIMESTAMP:
Deciphering the meaning of 42........

100 Largest Transactions  (last 24h):   450,019 BTC     ($8,519,080,617)    24.05% Total

Hashrate	 =  219.698 Ehash/s  (+8.30% in 24 hours)

Difficulty	=  32.045 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 756000   (in 453 blocks ~ 3 days 5 hours)

"The design supports a tremendous variety of possible transaction types, that I designed years ago.  Escrow transactions, bonded contracts, third-party arbitration, multi-party signature's etc...  All of these.. had to be designed in the beginning, to make sure they would be possible later."

💙⚡💜