⌛ lnd: 11:11p.m ET ⚡ October 01, 2022
The statistics are aggregated and calculated from multiple nodes within the Lightning Network. Due to the decentralized nature of the Lightning Network, the numbers observed are approximations and nodes that don't broadcast their state are NOT included***

(BTC) TImechain:
Block Count	=  756,656   (2022-10-01 23:11:35)

Hashrate =  273.527 Ehash/s   (+11.05% in 24 hours)
Difficulty	=  31.361 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 758016  (in 1358 blocks ~ 8 days 20 hours)

💙⚡💜