⌛ ₿lock 758234 ⚡ Next retarget @ block 758016
Github release	 v23.0   (2022-04-25)
Github stars	~  66459
Github last commit	~  2022-10-11

Reward Per Block = 6.25+0.0549 BTC ($120,154.54) 
    ~ next halving @ block 840000 (in 81766 blocks ~ 556 days) ~ 

 Hashrate  =  297.107 Ehash/s   (+17.11% in 24 hours)
 Difficulty   =  35.611 T   (+8.24% in 24 hours)

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 760032 (in 1798 blocks ~ 10 days 18 hours)


💙⚡💜   🖤⚡🖤