⌛ ₿lock 760841 ⚡ Next retarget @ block 762048
Github release	v23.0  (2022-04-25)
Github last commit	2022-10-29

Block Count	760,839 (2022-10-29 11:11:55)

Hashrate    =  288.19 Ehash/s  (+12.84% in 24 hours)
Difficulty  =  36.836 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 762048  (in 1206 blocks ~ 8 days 11 hours)

💙⚡💜