⌛ ₿lock 766214 ⚡ Next retarget @ block 768096
Github release	v24.0 (2022-11-24)

Reward Per Block	 =  6.25+0.1653 BTC ($109,491.79) 
next halving @ block 840000   (in 73786 blocks ~ 505 days)

Hashrate    =  254.189 Ehash/s (-0.61% in 24 hours)
Difficulty  =  34.304 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 768096 (in 1882 blocks ~ 13 days 14 hours)