⌛ ₿lock 768145 ⚡ Next retarget @ block 770112
Github release  v24.0.1    (2022-12-12)
Github last commit	~   2022-12-19

Reward Per Block = 6.25+0.1347 BTC ($104,895.04) 

Hashrate  =  254.318 Ehash/s  (+7.47% in 24 hours)
Difficulty  =  34.648 T  (+1.18% in 24 hours)

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 770112   (in 1967 blocks ~ 13 days 17 hours)


💙⚡💜