lnd⚡:  Block 770271 ⌛ Next retarget @ block 772128
Lightning (lnd⚡) Network Search & Analysis Engine:

The statistics are aggregated and calculated from multiple nodes within the Lightning Network. Due to the decentralized nature of the Lightning Network, the numbers observed are approximations and nodes that don't broadcast their state are NOT included***

Timechain Network STATS:
Block Time = 8min 50sec
Block Count	=  	770,271 (2023-01-03 22:11:26)

Reward Per Block	 = 6.25+0.0826 BTC  ($105,993.47) 

Difficulty   =  34.117 T 
Hashrate   =  257.491 Ehash/s  (-5.28% in 24 hours)

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 772128   (in 1857 blocks ~ 12 days 0 hours)

💙⚡💜