lnd: Lightning ⚡ Network Search & Analysis Engine
The statistics are aggregated and calculated from multiple nodes within the Lightning Network. Due to the decentralized nature of the Lightning Network, the numbers observed are approximations and nodes that don't broadcast their state are not included.

Connect to 1ML public node:

0217890e3aad8d35bc054f43acc00084b25229ecff0ab68debd82883ad65ee8266@23.237.77.11:9735

Bitcoin - Timechain STATS:
Github release	v24.0.1 (2022-12-12)
Github last commit	2023-01-06

Block Height = 771044

Reward Per Block  =  6.25+0.04421 BTC   ($107,833.96) 
   
Next halving @ block 840000   (in 68957 blocks ~ 472 days)

Hashrate   =  258.214 Ehash/s -14.35% in 24 hours
Difficulty   =  34.094 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 772128   (in 1084 blocks ~ 6 days 22 hours)

💙⚡💜