⌛ ₿lock 771818⚡ Next retarget @ block 772128
Github release	v24.0.1   (2022-12-12)
Last commit	 ~  2023-01-13

Reward Per Block	  =   6.25+0.1049 BTC   ($126,563.91) 

Hashrate  =   257.473 Ehash/s  (+3.6% in 24 hours)
Difficulty  =   34.094 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 772128  (in 310 blocks ~ 2 days 0 hours)

01100010 01101001 01100111 00100000 01110011 01110100 01101001 01100011 01101011 00100000 01100111 01110010 01100101 01100001 01110100 01100101 01110010 00100000 01110100 01101000 01100001 01101110 00100000 01101100 01101001 01110100 01110100 01101100 01100101 00100000 01110011 01110100 01101001 01100011 01101011

💙⚡💜