⌛ ₿lock 772129 ⚡ Next retarget @ block 774144
Github release	v24.0.1   (2022-12-12)
Last commit	  ~   2023-01-14

Reward Per Block   =   6.25+0.04998 BTC   ($131,717.02)

Hashrate    =  287.768 Ehash/s  (+4.14% in 24 hours)
Difficulty  =  34.136 T   (+0.12% in 24 hours)

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 774144   (in 2015 blocks ~ 12 days 18 hours)

💙⚡💜