⌛ ₿lock 773347 ⚡ Next retarget @ block 774144
Github release	v24.0.1    (2022-12-12)
Last commit	  ~   2023-01-23

Reward Per Block  =  6.25+0.09844 BTC   ($146,406.02)

Hashrate    =  288.237 Ehash/s    (+5.1% in 24 hours)
Difficulty  =  37.59 T 

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 774144   (in 797 blocks ~ 5 days 12 hours)

💙⚡💜