⌛ ₿lock 774220 ⚡ Next retarget @ block 776160
Github release	v24.0.1   (2022-12-12)
Last commit on  2023-01-28

Reward Per Block = 6.25+0.05395 BTC ($149,268.58)

Hashrate    =  311.037 Ehash/s   (+6.68% in 24 hours)
Difficulty  =  38.465 T    (+2.33% in 24 hours)

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 776160   (in 1940 blocks ~ 12 days 12 hours)

💙💜