⌛ ₿lock 775406 ⚡ Next retarget @ block 776160
Github release	v24.0.1   (2022-12-12)
Last commit on	~  2023-02-06

Reward Per Block  =  6.25+0.1403 BTC   ($146,584.88) 

Hashrate    =  278.355 Ehash/s   (-6.91% in 24 hours)
Difficulty  =  39.351 T

DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.

Next retarget @ block 776160   (in 754 blocks ~ 5 days 10 hours)

💙 💜