⌛ ₿lock 777044 ⚡ Next retarget @ block 778176
Github release  v24.0.1   (2022-12-12)
Last commit on   2023-02-17

1 BTC  =  24,555.42 USD   (2023-02-17 15:22:42)

Reward Per Block  =  6.25+0.179 BTC    ($157,761.59)

Hashrate  =    334.242 Ehash/s  (-0.45% in 24 hours)



Difficulty  =    39.156 T 


DIFFICULTY* is adjusted every 2016 blocks based on the time it took to find the previous 2016 blocks. At the desired rate of one block each 10 minutes, 2016 blocks would take exactly two weeks to find. If the previous 2016 blocks took more than two weeks to find, the difficulty is reduced. If they took less than two weeks, the difficulty is increased. The change in difficulty is in proportion to the amount of time over or under two weeks the previous 2016 blocks took to find. To find a block, the hash must be less than the target. The hash is effectively a random number between 0 and 2**256-1.


Next retarget @ block 778176 (in 1132 blocks ~ 7 days 1 hours)

💙 サトッ 💜